add memory catalog
