--- crates/integrations/datafusion/testdata/hms_catalog/core-site.xml
@@ -1,51 +1,0 @@
-<!--
-   Licensed to the Apache Software Foundation (ASF) under one or more
-   contributor license agreements.  See the NOTICE file distributed with
-   this work for additional information regarding copyright ownership.
-   The ASF licenses this file to You under the Apache License, Version 2.0
-   (the "License"); you may not use this file except in compliance with
-   the License.  You may obtain a copy of the License at
-
-       http://www.apache.org/licenses/LICENSE-2.0
-
-   Unless required by applicable law or agreed to in writing, software
-   distributed under the License is distributed on an "AS IS" BASIS,
-   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-   See the License for the specific language governing permissions and
-   limitations under the License.
--->
-
-<configuration>
-    <property>
-        <name>fs.defaultFS</name>
-        <value>s3a://warehouse/hive</value>
-    </property>
-    <property>
-        <name>fs.s3a.impl</name>
-        <value>org.apache.hadoop.fs.s3a.S3AFileSystem</value>
-    </property>
-    <property>
-        <name>fs.s3a.fast.upload</name>
-        <value>true</value>
-    </property>
-    <property>
-      <name>fs.s3a.endpoint</name>
-      <value>http://minio:9000</value>
-    </property>
-    <property>
-      <name>fs.s3a.access.key</name>
-      <value>admin</value>
-    </property>
-    <property>
-      <name>fs.s3a.secret.key</name>
-      <value>password</value>
-    </property>
-    <property>
-      <name>fs.s3a.connection.ssl.enabled</name>
-      <value>false</value>
-    </property>
-    <property>
-      <name>fs.s3a.path.style.access</name>
-      <value>true</value>
-    </property>
-</configuration>
